alx_travel_app/requirements.txt
asgiref==3.9.0
Django==5.2.4
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.14.1
django-cors-headers==4.7.0
django-environ==0.12.0
drf-yasg==1.21.10
mysqlclient==2.2.7
celery==5.5.3